requirements/requirements.txt
Django==1.9.4
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
PyJWT==1.4.0
requests==2.10.0
